requirements.txt
pydub
srt
pywav
pydub
google-cloud-storage
google-cloud-speech
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
google-api-core==1.16.0
google-auth==1.13.1
google-cloud-speech==1.3.2
googleapis-common-protos==1.51.0
grpcio==1.28.1
idna==2.9
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.23.1
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
srt==3.0.0
urllib3==1.25.8

